+++ App.tsx
@@ -3,12 +3,10 @@ import {Provider} from 'react-redux';
 import {store} from './src/store/store';
 import {MainNavigator} from './src/navigation/main';
 
-function App(): React.JSX.Element {
+export const App: React.FC = () => {
   return (
     <Provider store={store}>
       <MainNavigator />
     </Provider>
   );
-}
-
-export default App;
+};
