+++ package.json
@@ -13,9 +13,7 @@
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/native-stack": "^6.11.0",
     "@reduxjs/toolkit": "^2.2.8",
-    "i18next": "^23.16.0",
     "react": "18.3.1",
-    "react-i18next": "^15.0.3",
     "react-native": "0.75.4",
     "react-native-localization": "^2.3.2",
     "react-native-safe-area-context": "^4.11.0",
