+++ src/navigation/main.tsx
@@ -5,7 +5,7 @@ import {LaunchScreen} from '../screens/LaunchScreen';
 import {OnboardingScreen} from '../screens/onboarding/OnboardingScreen';
 import {AccountScreen} from '../screens/auth/AccountScreen';
 import {RestaurantsScreen} from '../screens/main/restaurants/RestaurantsScreen';
-import {RootStackParamList, ScreenNames} from './types';
+import {RootStackParamList, screenNames} from './types';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
@@ -13,14 +13,14 @@ export const MainNavigator = () => {
   return (
     <NavigationContainer>
       <Stack.Navigator>
-        <Stack.Screen name={ScreenNames.Launch} component={LaunchScreen} />
+        <Stack.Screen name={screenNames.launch} component={LaunchScreen} />
         <Stack.Screen
-          name={ScreenNames.Onboarding}
+          name={screenNames.onboarding}
           component={OnboardingScreen}
         />
-        <Stack.Screen name={ScreenNames.Account} component={AccountScreen} />
+        <Stack.Screen name={screenNames.account} component={AccountScreen} />
         <Stack.Screen
-          name={ScreenNames.Restaurants}
+          name={screenNames.restaurants}
           component={RestaurantsScreen}
         />
       </Stack.Navigator>
