--- src/utilities/localization/en-US.ts
@@ -1,6 +1,0 @@
-export default {
-  how: 'How do you want your egg today?',
-  boiledEgg: 'Boiled egg',
-  softBoiledEgg: 'Soft-boiled egg',
-  choice: 'How to choose the egg',
-};
